requirements.txt
azure-functions==1.20.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3